+++ tests/testthat/test-series.R
@@ -422,12 +422,12 @@ test_that("rep", {
 
 test_that("Series list", {
   series_list <- pl$DataFrame(list(a = c(1:5, NA_integer_)))$select(
-    pl$col("a")$list()$list()$append(
+    pl$col("a")$implode()$implode()$append(
       (
-        pl$col("a")$head(2)$list()$append(
-          pl$col("a")$tail(1)$list()
+        pl$col("a")$head(2)$implode()$append(
+          pl$col("a")$tail(1)$implode()
         )
-      )$list()
+      )$implode()
     )
   )$get_column("a") # get series from DataFrame
 
@@ -525,3 +525,10 @@ patrick::with_parameters_test_that("mean, median, std, var", {
   },
   .cases = make_cases()
 )
+
+
+test_that("n_unique", {
+  x = c(1:4, NA, NaN, 1) #6 unique one repeated
+  expect_identical(pl$Series(x)$n_unique(),6)
+  expect_grepl_error(pl$Series(c())$n_unique(),"operation not supported for dtype")
+})