+++ NEWS.md
@@ -1,5 +1,9 @@
 # polars (development version)
 
+## What's changed
+ - lazy functions translated: `pl$implode`, `pl$explode`, `pl$unique`, `pl$approx_unique`, `pl$head`, `pl$tail` (#196)
+ - `pl$list` is deprecated, use `pl$implode` instead (#196) 
+
 # polars 0.6.0
 
 ## BREAKING CHANGES